+++ src/firmware/drivers/hmc5883l_mag/hmc5883l_mag.h
@@ -1,5 +1,7 @@
 // hmc5883l_mag.h
 
+#include <stdint.h>
+
 #define HMC5883L_WRITE 0x3C
 #define HMC5883L_READ 0x3D
 
@@ -7,6 +9,6 @@ void HMC5883L_init(void);
 void HMC5883L_Update(void);
 
 // Access Functions
-float HMC5883L_GetX(void);
-float HMC5883L_GetY(void);
-float HMC5883L_GetZ(void);
+int16_t HMC5883L_GetX(void);
+int16_t HMC5883L_GetY(void);
+int16_t HMC5883L_GetZ(void);
